refactor(xlsx): extract chart module from parser lib.rs
Move chart XML parsing (~1550 lines) into parser/src/chart.rs: chart
anchors, chartSpace parsing, axis/title extractors, series/data-label/
error-bar parsing, and the DrawingML color helpers used only by charts.
resolve_fill_color stays in lib.rs (shared with drawing) and is now
pub(crate).

lib.rs drops from 4220 to 2674 lines. Verified behavior-invariant via
VRT regression mode (65/65 pass, 0.0% diff).
